build: check point
I believe this may fix many of the the outstanding issues maintaining this project

most of messy status when I have not recently run just clean
DirectoryIndex component
styles in dev versus prod mode
duplicate dumbledore pages

+++ packages/assets/bin/clean.sh
@@ -14,7 +14,7 @@ else
 fi
 
 # Check assets/filescreated directory
-if [ -d "$PWD"/assets/filescreated ] && [ "$(ls -A "$PWD"/assets/filescreated/*.txt 2> /dev/null)" ]; then
+if [ -d "$PWD"/assets ] && [ -d "$PWD"/assets/filescreated ] && [ "$(ls -A "$PWD"/assets/filescreated/*.txt 2> /dev/null)" ]; then
   for f in "$PWD"/assets/filescreated/*.txt; do
     echo "cleaning files listed in $f"
     cat "$f" | xargs -I{} rm "{}"
